Digital Scale With Camera(Project).rar
The project involves developing an advanced digital scale that can recognize three regional fruits (banana, passion fruit, and tangerine) using an integrated camera module. This scale will be utilized to assemble fruit baskets weighing up to 2000g, with individual packaging of 500g. The system controls product pricing, with a maximum value of €5 and an additional tax calculated based on a specific criterion related to the oldest student's number in the group.

The project's circuitry includes inputs from the camera module, the weight of the selected product, and an end-of-purchase button. Its outputs consist of the total amount due, receipt issuance, and the applied tax. The development requires proficiency in Verilog and electrical schematics, avoiding the use of standard memory components from the ISE software of Xilinx.

The system will be integrated with previous solutions if necessary, and the final submission includes a detailed report. The project will be evaluated based on criteria such as solution choice, efficient hardware use, report quality, solution robustness, and adherence to deadlines.